update based on comments to PR-1
