--- .editorconfig
@@ -1,9 +1,0 @@
-root = true
-
-[*]
-charset = utf-8
-
-[*.java]
-indent_style = space
-indent_size = 4
-trim_trailing_whitespace = true
