+++ .gitignore
@@ -2,3 +2,4 @@
 /.settings/
 /target/
 /.idea/
+/.editorconfig
