--- src/main/java/com/yourcodereview/java01/EugeneT1024/AtmCalculator.java
@@ -1,9 +1,0 @@
-package com.yourcodereview.java01.EugeneT1024;
-
-import java.util.List;
-
-public interface AtmCalculator {
-
-    CalculationResult calculate(int cashAmount, List<Integer> denominationsOfBanknotes);
-
-}
